+++ src/managednetworkfabric/azext_managednetworkfabric/tests/latest/test_tap_resync.py
@@ -35,7 +35,7 @@ def step_resync(test, checks=None):
     if checks is None:
         checks = []
     test.cmd(
-        'az networkfabric tap resync --resource-group {rg} --network-tap-name {name}', checks=checks)
+        'az networkfabric tap resync --resource-group {rg} --resource-name {name}', checks=checks)
 
 
 class GA_TapResyncScenarioTest1(ScenarioTest):