+++ src/managednetworkfabric/HISTORY.rst
@@ -2,6 +2,11 @@
 
 Release History
 ===============
+6.1.0
+++++++
+* Added device update-admin-state support
+* Changed network-tap-name parameter in tap resync to resource-name
+
 6.0.0
 ++++++
 * Added device run rw support